requirements.txt
asgiref==3.7.2
Django==3.2.3
djangorestframework==3.14.0
Pillow==10.1.0
pytz==2023.3.post1
sqlparse==0.4.4
